Adding Reflection report here too, as it may be easier to read
This semester is my first year of my degree, and as this has been one of my first classes in the faculty of Engineering and Information Technology, it has been interesting. For my groups report, my contribution was the background research and context. I found this interesting as I enjoyed learning about how to roll out an effective vaccine. It also helped my broader knowledge in the area as I did not know about the complexities involved in the strategies for vaccine implementation. I also enjoyed this as doing the more technical aspects of the assignment would have not been my strongpoint.  

 

The rest of the assignment was split between my fellow group members. Denise was in charge of the introduction of our report and explaining the 3 models we had selected to look at our problem: Agent Based Models, Alternate Path Models, and Cheapest Network Model. Vinh identified the current models being used and discussed how they did not fully address or solve the problem. And finally, Arjun described and illustrated the data sources for our model and explained how each of the 3 models will be applied to our problem.  

The majority of the coding element of the assignment was completed by Arjun and Vinh, as they decided they both wanted to do the code only option. They did endure the most of it, while Denise and I offered support where we could.  

 

The positives I learnt from this project were: teamwork seriously does make the dream work, the knowledge I gained about the concept of modelling and …. |    Our team was at a serious disadvantage at the start, because it took us ages to get organized and catch up as a group. This was due to many varied reasons, such as Vinh conducting his learning from distance, and the rest of us not having any contact with each other apart from our time from the Zooms. However, through teamwork and communication we passed this adversity, fairly divvying up the work. This landed us an exceedingly high mark, in our pre-submission. This showed me how effective teamwork is, when used. Despite the negative outcomes we faced. The knowledge I gained on modelling from this assignment was hugely positive. While most of the knowledge I have, has been learnt from the lectures. This consolidated it further, allowing me to put it into practice.  

 

 

If I was faced with a real-world problem, and had to use modelling to solve it, I would start by analyzing the problem, by researching its background and context, just like I have done in this assignment.  This would allow me to ascertain the complexities of the problem beforehand. Which is vital when deciding what models I would use to solve them because the implementation of the most effective model depends on the problem, and if I didn’t understand the problem, it will be impossible to implement the most effective solution. For example, if the problem is complex with many different contributing non-related factors, I would use Agent Based Modelling, as it allows the thinker to compare the different traits of the thing being modelled. However, if the problem is not as complex and the cause of the problem is simpler, a model such as the Alternate Paths Model, where only one trait is considered in the model. I then would make a prototypes and collect data.